update dwg object writer module 2025-06-04  MerlotRain  <merlotrain@gmail.com>
    * include/dwg/xdata/ExtendedDataDictionary.h:
    * src/io/dwg/writers/DwgObjectWriter_Common.cpp:
    * src/io/dwg/writers/DwgObjectWriter_Entities.cpp:

+++ src/io/dwg/writers/DwgObjectWriter_Entities.cpp
@@ -448,7 +448,7 @@ void DwgObjectWriter::writeFace3D(Face3D *face)
     //R2000 +:
     if (R2000Plus)
     {
-        bool noFlags = face->flags() == InvisibleEdgeFlag::None;
+        bool noFlags = face->flags() == (int) InvisibleEdgeFlag::None;
         //Has no flag ind. B
         _writer->writeBit(noFlags);
